thread.c: fix -Wformat warning
../thread.c:1219:18: warning: format specifies type 'unsigned long' but the argument has type 'rb_hrtime_t'
      (aka 'unsigned long long') [-Wformat]
                 end, now);
                 ^~~
../thread.c:1219:23: warning: format specifies type 'unsigned long' but the argument has type 'rb_hrtime_t'
      (aka 'unsigned long long') [-Wformat]
                 end, now);
                      ^~~

git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@64712 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ thread.c
@@ -1215,8 +1215,8 @@ hrtime_update_expire(rb_hrtime_t *timeout, const rb_hrtime_t end)
 
     if (now > end) return 1;
     thread_debug("hrtime_update_expire: "
-		 "%"PRI_64_PREFIX"u > %"PRI_64_PREFIX"u\n",
-                 end, now);
+                 "%"PRI_64_PREFIX"u > %"PRI_64_PREFIX"u\n",
+                 (unsigned long)end, (unsigned long)now);
     *timeout = end - now;
     return 0;
 }
